Make CIME_REL_DIR a cache variable
By changing to a cache variable it ensures that it is always set to
as sane default and that there is some tracability if it is changed.

Since this largely will be changed by the host code a simple command
can set and report where it was set from. Instead of the previous

set(CIME_REL_DIR <dir>)

it is now

set(CIME_REL_DIR <dir> CACHE PATH "set by the host code for reasons.")

The default was set to "." to maintain the current behavior.

Error checking now changes from is this variable set to is the variable
set to a directory that exists.