Update netcdf and hdf5 searching
Made the variable expected to be passed into cime configure a cache variable to
ensure proper tracibility.

Update search for hdf5 and netcdf to use find_path instead of if exists. This
might need to be expanded to use an actual find module to be very safe, but
for now this makes the search and usage to be more in line with cmake standards.