Make PIO an external project that is configured and built along with CIME
This uses the PIO that already exists in CIME instead of pointing to an
external TPL of PIO. The configure for PIO is done based on information
given to CIME and then a precache file is created from
pio_config.cmake.in to get the configuration to PIO.

This did require touching a number of the components so that they
point to the correct location for PIO now. Previously this was
given to nearly all targets as the include directories was being
set early and near the top of the directory hierarchy.